requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ecdsa==0.14.1
model-mommy==2.0.0
Pillow==8.1.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-jose==3.2.0
pytz==2021.1
rsa==4.7.1
six==1.15.0
sqlparse==0.4.1
